my_finance/requirements.txt
fastapi
pydantic
uvicorn
sqlalchemy
python-dotenv
pydantic[dotenv]
python-jose
passlib[bcrypt]
python-multipart